Add arch_variants field to tag_packages
The arch_variants field represents different build configurations that
should be used when building the package. If unset, only the base
configuration will be used. Otherwise, it should be set to a space-separated
list of variants with 'MAIN' representing the base configuration. For example
'MAIN app' might result in builds being build for x86_64 x86_64~app i686 i686~app.